Added Moduliths and created initial structure for ERP system
This commit introduces Moduliths to enhance modularity in the ERP application. This change affected the back-end part of the application, where various new files and tests were created. These tests are intended to verify the conformity of the application to the Moduliths structure. Also, new entity classes were added to model the various components of the ERP system.