Refactor prefecture and city classes and remove unused tests
This commit introduces a significant refactoring to the prefecture and city classes, in addition to related classes. The first action was to refactor the prefecture attribute class to use PrefectureNames instead of PrefectureName. Besides, the Name.java test file was deleted as seemed to no longer be used, and the creation of new tests such as CityRecordTest.java and CityLgCodeTest.java has been done. Furthermore, some renaming has been done for clarity and consistency purposes. For example, the NameKanaTest.java was renamed to PrefectureKanaTest.java, mirroring the structure of the entities under test. Also, new classes were created, like CountyNamesTest.java and CountryRomaTest.java. The changes in this commit are meant to improve readability, maintainability, and overall code quality.